@@ -40,18 +40,18 @@
   "devDependencies": {
     "@babel/core": "^7.11.6",
     "@fortawesome/free-brands-svg-icons": ">=5.15.3",
-    "@storybook/addon-a11y": "^6.5.9",
-    "@storybook/addon-actions": "^6.5.9",
-    "@storybook/addon-docs": "^6.5.9",
+    "@storybook/addon-a11y": "^6.5.10",
+    "@storybook/addon-actions": "^6.5.10",
+    "@storybook/addon-docs": "^6.5.10",
     "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/addon-storysource": "^6.5.9",
-    "@storybook/addon-toolbars": "^6.5.9",
-    "@storybook/addon-viewport": "^6.5.9",
-    "@storybook/addons": "^6.5.9",
-    "@storybook/builder-vite": "^0.1.38",
-    "@storybook/react": "^6.5.9",
+    "@storybook/addon-storysource": "^6.5.10",
+    "@storybook/addon-toolbars": "^6.5.10",
+    "@storybook/addon-viewport": "^6.5.10",
+    "@storybook/addons": "^6.5.10",
+    "@storybook/builder-vite": "^0.2.2",
+    "@storybook/react": "^6.5.10",
     "@storybook/testing-library": "^0.0.13",
-    "@storybook/theming": "^6.5.9",
+    "@storybook/theming": "^6.5.10",
     "@testing-library/dom": "^8.13.0",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.2.0",
@@ -59,7 +59,7 @@
     "@types/node": "^12.0.0",
     "@typescript-eslint/eslint-plugin": "^4.4.0",
     "@typescript-eslint/parser": "^4.4.0",
-    "@vitejs/plugin-react": "^1.3.2",
+    "@vitejs/plugin-react": "^2.0.1",
     "auto": "^10.32.6",
     "autoprefixer": "^9.7.2",
     "babel-eslint": "10.1.0",
@@ -74,7 +74,7 @@
     "eslint-plugin-jsx-a11y": "6.3.1",
     "eslint-plugin-react": "7.20.6",
     "eslint-plugin-react-hooks": "4.1.2",
-    "eslint-plugin-storybook": "^0.5.13",
+    "eslint-plugin-storybook": "^0.6.4",
     "husky": "^3.0.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^28.1.1",
@@ -88,10 +88,10 @@
     "react-docgen-typescript-loader": "^3.1.0",
     "ts-jest": "^28.0.5",
     "tslib": "2.3.1",
-    "typescript": "4.7.3",
-    "vite": "^2.9.13",
-    "vite-plugin-css-injected-by-js": "^1.5.1",
-    "vite-plugin-svgr": "^2.2.0"
+    "typescript": "4.8.2",
+    "vite": "^3.0.9",
+    "vite-plugin-css-injected-by-js": "^2.0.4",
+    "vite-plugin-svgr": "^2.2.1"
   },
   "optionalDependencies": {
     "fsevents": "^2.2.1"
@@ -111,7 +111,8 @@
     "@babel/core": "7.13.15",
     "@types/react": "18.0.14",
     "@types/react-dom": "18.0.5",
-    "tslib": "2.3.1"
+    "tslib": "2.3.1",
+    "ts-node": "10.9.1"
   },
   "browserslist": [
     ">0.2%",
